refactor: duplicated code
