+++ packages/cli/src/commands/deploy.ts
@@ -1,31 +1,21 @@
-import fs from 'node:fs';
 import path from 'node:path';
 import { createDeployment, createFunction, getDeploymentConfig, writeDeploymentConfig } from '../utils/deployments';
-import { logInfo, logError, logSuccess, logSpace } from '../utils/logger';
+import { logInfo, logSpace, logDeploymentSuccessful } from '../utils/logger';
 import inquirer from 'inquirer';
-import { SUPPORTED_EXTENSIONS } from '../utils/constants';
 import { authToken } from '../auth';
 import { trpc } from '../trpc';
-import chalk from 'chalk';
+import { getAssetsDir, getFileToDeploy } from '../utils';
 
 export async function deploy(file: string, { preact, publicDir }: { preact: boolean; publicDir: string }) {
-  const fileToDeploy = path.join(process.cwd(), file);
-  const assetsDir = path.join(path.parse(fileToDeploy).dir, publicDir);
+  const fileToDeploy = getFileToDeploy(file);
 
-  if (!fs.existsSync(fileToDeploy) || fs.statSync(fileToDeploy).isDirectory()) {
-    logError(`File '${fileToDeploy}' does not exists/is not a file.`);
+  if (!fileToDeploy) {
     return;
   }
 
-  const extension = path.extname(fileToDeploy);
+  const assetsDir = getAssetsDir(fileToDeploy, publicDir);
 
-  if (!SUPPORTED_EXTENSIONS.includes(extension)) {
-    logError(`Extension '${extension}' is not supported (${SUPPORTED_EXTENSIONS.join(', ')})`);
-    return;
-  }
-
-  if ((!fs.existsSync(assetsDir) || !fs.statSync(assetsDir).isDirectory()) && publicDir !== 'public') {
-    logError(`Public directory '${publicDir}' does not exist.`);
+  if (!assetsDir) {
     return;
   }
 
@@ -70,16 +60,11 @@ export async function deploy(file: string, { preact, publicDir }: { preact: bool
         },
       ]);
 
+      logSpace();
+
       const deployment = await createDeployment(func, fileToDeploy, preact, assetsDir);
       writeDeploymentConfig(fileToDeploy, { functionId: func, organizationId: organization, publicDir });
-
-      logSpace();
-      logSuccess(`Function deployed.`);
-      logSpace();
-      console.log(
-        ` ➤ ${chalk.gray('https://') + chalk.blueBright(deployment.functionName) + chalk.gray('.lagon.app')}`,
-      );
-      logSpace();
+      logDeploymentSuccessful(false, deployment.functionName);
     } else {
       const { name } = await inquirer.prompt([
         {
@@ -90,14 +75,10 @@ export async function deploy(file: string, { preact, publicDir }: { preact: bool
       ]);
 
       logSpace();
+
       const func = await createFunction(name, fileToDeploy, preact, assetsDir);
       writeDeploymentConfig(fileToDeploy, { functionId: func.id, organizationId: organization, publicDir });
-
-      logSpace();
-      logSuccess(`Function ${name} created.`);
-      logSpace();
-      console.log(` ➤ ${chalk.gray('https://') + chalk.blueBright(name) + chalk.gray('.lagon.app')}`);
-      logSpace();
+      logDeploymentSuccessful(true, name);
     }
 
     return;
@@ -112,9 +93,5 @@ export async function deploy(file: string, { preact, publicDir }: { preact: bool
   }
 
   const deployment = await createDeployment(config.functionId, fileToDeploy, preact, finalAssetsDir);
-  logSpace();
-  logSuccess(`Function deployed.`);
-  logSpace();
-  console.log(` ➤ ${chalk.gray('https://') + chalk.blueBright(deployment.functionName) + chalk.gray('.lagon.app')}`);
-  logSpace();
+  logDeploymentSuccessful(false, deployment.functionName);
 }
